+++ src/Forms/CreateVinylCassette.cs
@@ -9,18 +9,30 @@ namespace Cassiopeia.src.Forms
         private AlbumData album;
         private VinylAlbum creatingVinyl = null;
         private VinylAlbum editingVinyl;
+        private CassetteTape tape;
         private int numDisc;
         private int NCA, NCB;
         private bool edit = false;
         private char side;
         private bool canceled = true;
-        public CreateVinylCassette(ref AlbumData a, bool vinyl = true)
+        private bool cassette = false;
+        public CreateVinylCassette(ref AlbumData a, bool cassette)
         {
             InitializeComponent();
+            this.cassette = cassette;
             album = a;
-            Log.Instance.PrintMessage("Creating a Vinyl. Album length: " + a.Length, MessageType.Info);
-            numDisc = 1;
-            side = (char)('A' + (numDisc + (numDisc - 2)));
+            if(!cassette)
+            {
+                Log.Instance.PrintMessage("Creating a Vinyl. Album length: " + a.Length, MessageType.Info);
+                numDisc = 1;
+                side = (char)('A' + (numDisc + (numDisc - 2)));
+            }
+            else
+            {
+                Log.Instance.PrintMessage("Creating a Cassette Tape. Album length: " + a.Length, MessageType.Info);
+                numDisc = 1;
+                side = 'A';
+            }
             //Now we need to check if the album needs one or more Vinyl.
             SetMaxLength();
             PutTexts();
@@ -61,7 +73,9 @@ namespace Cassiopeia.src.Forms
 
         private void PutTexts()
         {
-            if (creatingVinyl is not null || numDisc == 1)
+            if (cassette)
+                Text = Kernel.GetText("creando") + " " + Kernel.GetText("tape");
+            else if (creatingVinyl is not null || numDisc == 1)
                 Text = Kernel.GetText("creando") + " " + Kernel.GetText("vinyl") + " " + numDisc;
             else
                 Text = Kernel.GetText("editando") + " " + Kernel.GetText("vinyl") + " " + numDisc;
@@ -137,7 +151,7 @@ namespace Cassiopeia.src.Forms
             //}
 
         }
-        private void CreateNewVinyl(int nsFront, int nsBack)
+        private void CreateRecord(int nsFront, int nsBack)
         {
             MediaCondition exterior = (MediaCondition)Enum.Parse(typeof(MediaCondition), comboBoxEstadoExterior.SelectedIndex.ToString());
             MediaCondition medio = (MediaCondition)Enum.Parse(typeof(MediaCondition), comboBoxEstadoMedio.SelectedIndex.ToString());
@@ -146,7 +160,10 @@ namespace Cassiopeia.src.Forms
             //Creating Vinyl
             try
             {
-                creatingVinyl = new VinylAlbum(album, nsFront, nsBack, exterior, medio, Convert.ToInt16(textBoxAño.Text), textBoxPais.Text);
+                if (!cassette)
+                    creatingVinyl = new VinylAlbum(album, nsFront, nsBack, exterior, medio, Convert.ToInt16(textBoxAño.Text), textBoxPais.Text);
+                else
+                    tape = new CassetteTape(album, nsFront, nsBack, exterior, medio, Convert.ToInt16(textBoxAño.Text), textBoxPais.Text);
             }
             catch (Exception)
             {
@@ -154,8 +171,16 @@ namespace Cassiopeia.src.Forms
                 MessageBox.Show(Kernel.GetText("errorAño"));
                 throw;
             }
-            Kernel.Collection.AddVinyl(ref creatingVinyl);
-            Log.Instance.PrintMessage("Vinyl added OK", MessageType.Correct);
+            if(!cassette)
+            {
+                Kernel.Collection.AddVinyl(ref creatingVinyl);
+                Log.Instance.PrintMessage("Vinyl added OK", MessageType.Correct);
+            }
+            else
+            {
+                Kernel.Collection.AddTape(ref tape);
+                Log.Instance.PrintMessage("Tape added OK", MessageType.Correct);
+            }
             Kernel.SetSaveMark();
         }
         private void buttonOK_Click(object sender, EventArgs e)
@@ -197,14 +222,14 @@ namespace Cassiopeia.src.Forms
                 //We create the Vinyl
                 try
                 {
-                    CreateNewVinyl(NCA, NCB);
+                    CreateRecord(NCA, NCB);
                 }
                 catch (Exception)
                 {
                     return;
                 }
                 //The user might want a smaller CD, excluding the bonus songs. Or we need another disc 
-                if (NCA + NCB != album.NumberOfSongs)
+                if (NCA + NCB != album.NumberOfSongs && !cassette)
                 {
                     AnotherDisc();
                 }