+++ src/Log.cs
@@ -30,9 +30,16 @@ namespace Cassiopeia
         {
             timeSinceStart = Stopwatch.StartNew();
             logView = new LogViewer();
-            file = new StreamWriter(Environment.CurrentDirectory + "\\log.txt", false);
-            file.AutoFlush = true;
-            logView = new LogViewer();
+            try
+            {
+                file = new StreamWriter(Environment.CurrentDirectory + "\\log.txt", false);
+                file.AutoFlush = true;
+                logView = new LogViewer();
+            }
+            catch (Exception)
+            {
+                throw;
+            }
             if (file != null)
             {
                 PrintInfo();