+++ changelog.md
@@ -1,4 +1,8 @@
 # Cassiopeia Changelog
+#### 2.0.247.0
+- Cassette tapes
+- Cassiopeia exists normally if it finds another instance running [FIX]
+
 #### 2.0.246.0
 - Added a button to share lyrics on Twitter